Formatting

+++ src/main/scala/com/example/DspFrontend.scala
@@ -2,21 +2,14 @@ package com.example
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.model.StatusCodes
-import akka.http.scaladsl.server.Directives
-import akka.http.scaladsl.server.Route
+import akka.http.scaladsl.server.{Directives, Route}
 import akka.stream.Materializer
 
-import scala.concurrent.ExecutionContextExecutor
 import scala.util.{Failure, Success}
 
-/*
- * Add your logic here. Feel free to rearrange the code as you see fit,
- * this is just a starting point.
- */
 trait DspFrontend extends Directives with BidResponseJsonFormats {
 
   implicit val system: ActorSystem
-//  implicit val executor: ExecutionContextExecutor
   implicit val materializer: Materializer
 
   var service: CampaignService
@@ -33,7 +26,7 @@ trait DspFrontend extends Directives with BidResponseJsonFormats {
     } ~ path("winner" / Segment) { auction_id =>
           onComplete(service.bidWinner(auction_id)) {
             case Success(_) => complete(StatusCodes.Created, "OK")
-            case Failure(_) => complete(StatusCodes.InternalServerError.intValue, "No entry found for given auction token. Bad Request or Possible Timeout")
+            case Failure(_) => complete(StatusCodes.InternalServerError, "No entry found for given auction token. Bad Request or Possible Timeout")
       }
     }
 }
